excerpt post reaction as individual component
